requirements.txt
torch>=2.0
torchaudio>=2.0
pyyaml>=6.0
pytest>=7.0
